(refactor) refactor virtual terminal func

+++ cmd/web/render.go
@@ -27,7 +27,7 @@ var functions = template.FuncMap{
 }
 
 func formatCurr(n int) string {
-	f := float32(n / 100)
+	f := float32(n) / float32(100)
 	return fmt.Sprintf("$%.2f", f)
 }
